Bug fixed
Fixed a bug, origin being the 'mouse-following-tooltip', where if you hover (or touch in mobile devices) towards the right side of the screen, the tooltip follows you there (invisible) and expands width of the page (distorting it in the process) to take up its own space. Now it only moves when you hover over its direct parent (earlier it was on document 'mousemove') and also will move until space (width) of page available.